backend/cloud_functions/parse_v3/requirements.txt
cloudevents==1.*
google-events==0.14.*
pandas==2.2.3
functions-framework==3.*
google-cloud-core==2.4.*
google-cloud-firestore==2.19.*
google-cloud-storage==2.19.*
flask==2.0.2
Werkzeug==2.2.2
pyarrow==18.*
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.27.0
requests==2.32.0
db-dtypes==1.3.1